% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_intersect.R
\name{curve_intersect}
\alias{curve_intersect}
\title{Calculate the intersection of two curves}
\usage{
curve_intersect(curve1, curve2)
}
\arguments{
\item{curve1}{\code{data.frame} with columns named \code{x} and \code{y}}

\item{curve2}{\code{data.frame} with columns named \code{x} and \code{y}}
}
\description{
Calculate where two lines or curves intersect. Curves are defined as data
frames with x and y columns providing cartesian coordinates for the lines.
This function works on both linear and nonlinear curves.
}
\details{
For now, \code{curve_intersect} will only find one intersection.
}
\examples{
library(reconPlots)

# Straight lines
line1 <- data.frame(x = c(1, 9), y = c(1, 9))
line2 <- data.frame(x = c(9, 1), y = c(1, 9))

curve_intersect(line1, line2)

# Curved lines
curve1 <- data.frame(Hmisc::bezier(c(1, 8, 9), c(1, 5, 9)))
curve2 <- data.frame(Hmisc::bezier(c(1, 3, 9), c(9, 3, 1)))

curve_intersect(curve1, curve2)
}
